chatbot_ui/requirements.txt
streamlit>=1.32.0
google-cloud-aiplatform>=1.75.0
google-genai>=0.1.0
python-dotenv>=1.0.0

# Fix numpy compatibility
numpy>=1.26.0,<2.0.0

# Reinstall these after numpy
pandas>=2.0.0
pyarrow>=14.0.0
